move main testing setup to its own file
